Refactor: pledge page 내 CandidateSection 컴포넌트 통합
